+++ Packages/com.unity.inputsystem/InputSystem/Runtime/Actions/InputAction.cs
@@ -1274,9 +1274,11 @@ namespace UnityEngine.InputSystem
         public unsafe bool WasPressedThisFrame()
         {
             var state = GetOrCreateActionMap().m_State;
-            if (state != null && !state.IsSuppressed)
+            if (state != null)
             {
                 var actionStatePtr = &state.actionStates[m_ActionIndexInState];
+                if (actionStatePtr->isSuppressed)
+                    return false;
                 var currentUpdateStep = InputUpdate.s_UpdateStepCount;
                 return actionStatePtr->pressedInUpdate == currentUpdateStep && currentUpdateStep != default;
             }
@@ -1319,6 +1321,8 @@ namespace UnityEngine.InputSystem
             if (state != null)
             {
                 var actionStatePtr = &state.actionStates[m_ActionIndexInState];
+                if (actionStatePtr->isSuppressed)
+                    return false;
                 return actionStatePtr->framePressed == ExpectedFrame();
             }
 
@@ -1377,6 +1381,8 @@ namespace UnityEngine.InputSystem
             if (state != null)
             {
                 var actionStatePtr = &state.actionStates[m_ActionIndexInState];
+                if (actionStatePtr->isSuppressed)
+                    return false;
                 var currentUpdateStep = InputUpdate.s_UpdateStepCount;
                 return actionStatePtr->releasedInUpdate == currentUpdateStep && currentUpdateStep != default;
             }
@@ -1421,6 +1427,8 @@ namespace UnityEngine.InputSystem
             if (state != null)
             {
                 var actionStatePtr = &state.actionStates[m_ActionIndexInState];
+                if (actionStatePtr->isSuppressed)
+                    return false;
                 return actionStatePtr->frameReleased == ExpectedFrame();
             }
 
@@ -1478,9 +1486,11 @@ namespace UnityEngine.InputSystem
         {
             var state = GetOrCreateActionMap().m_State;
 
-            if (state != null && !state.IsSuppressed)
+            if (state != null)
             {
                 var actionStatePtr = &state.actionStates[m_ActionIndexInState];
+                if (actionStatePtr->isSuppressed)
+                    return false;
                 var currentUpdateStep = InputUpdate.s_UpdateStepCount;
                 return actionStatePtr->lastPerformedInUpdate == currentUpdateStep && currentUpdateStep != default;
             }
@@ -1520,6 +1530,8 @@ namespace UnityEngine.InputSystem
             if (state != null)
             {
                 var actionStatePtr = &state.actionStates[m_ActionIndexInState];
+                if (actionStatePtr->isSuppressed)
+                    return false;
                 return actionStatePtr->framePerformed == ExpectedFrame();
             }
 
@@ -1601,6 +1613,8 @@ namespace UnityEngine.InputSystem
             if (state != null)
             {
                 var actionStatePtr = &state.actionStates[m_ActionIndexInState];
+                if (actionStatePtr->isSuppressed)
+                    return false;
                 var currentUpdateStep = InputUpdate.s_UpdateStepCount;
                 return actionStatePtr->lastCompletedInUpdate == currentUpdateStep && currentUpdateStep != default;
             }
@@ -1642,6 +1656,8 @@ namespace UnityEngine.InputSystem
             if (state != null)
             {
                 var actionStatePtr = &state.actionStates[m_ActionIndexInState];
+                if (actionStatePtr->isSuppressed)
+                    return false;
                 return actionStatePtr->frameCompleted == ExpectedFrame();
             }