+++ Packages/com.unity.inputsystem/InputSystem/Runtime/Events/InputEventHandledPolicy.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace UnityEngine.InputSystem.LowLevel
 {
     /// <summary>
@@ -9,12 +11,18 @@ namespace UnityEngine.InputSystem.LowLevel
         /// <summary>
         /// Input events will be discarded directly and not propagate for state changes.
         /// </summary>
+        [Obsolete("Use SuppressActionEventNotifications instead. SuppressStateUpdates desynchronizes Input System state from source state, leading to undefined behavior.", error: false)]
         SuppressStateUpdates,
 
         /// <summary>
         /// Input events will be processed for state updates and input action interaction updates but interaction
         /// event notifications will be suppressed.
         /// </summary>
-        SuppressActionEventNotifications
+        SuppressActionEventNotifications,
+
+        /// <summary>
+        /// The default input event handling policy.
+        /// </summary>
+        Default = SuppressActionEventNotifications
     }
 }